updating scripts, scene, models and prefabs
Notes from DevLog:

GPS / Locations are working correctly, placed in the update method, they should be updating correctly. Need to test with moving around more significantly to make sure that values are changing properly. Store, Get, and Calculate distance buttons are all working as intended so far, need to make sure that the user can save over old coords and then get new ones after the first time.
I’ve changed the 3D mesh to be a map marker, but there is some troubleshooting to be done. First, the scale is too big. Scale it down. Second, I tried to attach a user input field to label the marker, at first it was attached to screen space, last build was world space, but I wasn’t able to see it. In addition, testing needs to be done on the save / load of this prefab. I have attached the Persistent Object scrip to the prefab, so in theory it should be working, but need to make sure.
Location / VPS implementations are partially there? I have a location scanned, but have been unsure how to work with it as intended. Planning to go back to tutorial tomorrow to make more progress with Lightship specific elements.